fix: implicit conversion from float to int warning (https://github.com/googleapis/gax-php/issues/438)

+++ Gax/src/PollingTrait.php
@@ -68,7 +68,7 @@ trait PollingTrait
             if ($pollCallable()) {
                 return true;
             }
-            $currentPollDelayMillis = min([
+            $currentPollDelayMillis = (int) min([
                 $currentPollDelayMillis * $pollDelayMultiplier,
                 $maxPollDelayMillis
             ]);
